fixed bug in continent filler
